✨feat: implement auto-login management on logout - 로그아웃 시 자동 로그인 정보 삭제 함수 추가 - 로그인 성공 시 자동 로그인 정보 저장 - 컴포넌트 마운트 시 자동 로그인 확인 및 처리 - LoginSignupPage/Login.tsx 파일 삭제 Related to: #174
